Work around some changes to Login flow in 3.8.9 and 4.0
In 3.8.9/4.0, we now need to hook the earlier callback, Session, in
order to avoid being skipped by the MaybeShowNoAuthPage method since
we're using a /NoAuth/Login.html page to collect login information.

This degrades nicely on 3.8.8 since the Session callback isn't called.

The Auth callback will get invoked on 3.8.9 but should bail quickly
because the user is logged in already.

git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/RT-Authen-ExternalAuth/trunk@20274 e417ac7c-1bcc-0310-8ffa-8f5827389a85