When I removed the 3.8.0 and 3.8.1 workaround, CanonicalizeUserInfo started running again
Before this, we weren't fetching external data from ldap or sql after
the auth.  This was spotted because the tests started failing (auth
worked but we couldn't create the user because it couldn't retrieve user
info for Create).

Defined an Info action to run.
apply some ldap screwing around so that filter is right and the
attributes line up and then try to forget how it all works.