Ensure that $session->{'CurrentUser'} always contains a value
Some logic in RT assumes that the session always contains a CurrentUser,
even if that user is empty.  Replace all instances of delete with a set
to a blank CurrentUser object.

This resolves CVE-2012-2770, wherein a RSS url for a user could be used
to gain access to a fully logged-in session as that user.