Add option to prepend values on LDAP search filter
Customer AD configurations contain entries in the form:

    proxyAddresses: smtp:email@test.com

where AD puts the 'smtp:' prefix on the email address. This commit
adds a configuration option to pass in a value to be prepended
to the incoming email address in the LDAP filter. This allows
alternate email addresses listed in attributes like proxyAddresses
to match.

It sets smtp: as a default for proxyAddresses since it seems to be
the common case for AD.