Remove erroneous docs.
These were written on the docs-improvements branch and describe code
from the multiple-emails branch, which has not been merged.

Shipping these docs makes end users believe that this extension contains
code that it does not.

Backing these out to a separate branch.  Unfortunately, multiple-emails
doesn't have a PODified RT_SiteConfig.pm so storing it there is hard and
this merge will be really interesting.