Remove erroneous docs.
These were written on the docs-improvements branch and describe code
from the multiple-emails branch, which has not been merged.

Shipping these docs makes end users believe that this extension contains
code that it does not.

The docs have been moved to the tip of multiple-emails.  They will
re-enter when that branch is merged.