Avoid sending blank lines at the top of responses
Because of a trailing newline in this file, when RT served images from
the non-static-handler, it would inject a blank line at the top of the
file, breaking it.  This caused custom logos and charts from search
results to fail to render.

This was caused when eeb05b8f lost the return; at the end of init,
allowing the blank line to be printed.

This restores the return, with a helpful comment for the future and also
futureproofs by removing the blank lines.